http: limit maximum buffer size (#497)
git-svn-id: http://svn.code.sf.net/p/qmmp-dev/code/trunk/qmmp@8490 90c681e8-e032-0410-971d-27865f9a5e38

+++ src/plugins/Transports/http/httpstreamreader.h
@@ -43,12 +43,12 @@ class HTTPInputSource;
 struct HttpStreamData
 {
     char *buf;
-    long buf_fill;
+    size_t buf_fill;
     QString content_type;
     bool aborted;
     QHash <QString, QByteArray> header;
     bool icy_meta_data;
-    int icy_metaint;
+    size_t icy_metaint;
 };
 
 /*! @internal
@@ -101,11 +101,11 @@ private:
     HttpStreamData m_stream;
     QString m_url;
     QString m_userAgent;
-    int m_metacount;
+    size_t m_metacount;
     QString m_title;
     bool m_ready;
     bool m_meta_sent;
-    long m_buffer_size;
+    size_t m_buffer_size;
     QTextCodec *m_codec;
     DownloadThread *m_thread;
     HTTPInputSource *m_parent;
